create Makefile.PL in filegatherer phase
This allows other plugins to not just see that this file exists (at early
phases such as file munging, and generate files accordingly -- e.g.
[InstallGuide]), but allow other plugins to add content earlier (e.g.
Devel::CheckLib). The template will be evaluated at the normal time, once the
prereq list is finalized.

This change mirrors that in rjbs/Dist-Zilla#229, and live independently of
that, but must release first.