Do not try to check min_perl_version during prereq registration
The call to $self->min_perl_version in the default sub for eumm_version would
be unlikely to get something correct. This is because eumm_version ends up
being called during prereq registration. That means that whatever plugin might
set a perl version prereq may not have run yet.

The symptom of this is you end up losing the min_perl_version bits in the
generated Makefile.PL, because min_perl_version is called too early and ends
up being set to an empty string.

Specifically, I saw this with a number of distro where I have a line like "use
5.0100" in the code and I'm using AutoPrereqs.